now with target locks!
More bug fixes, mainly around the target locking
system--it was confusing to me since polar coords
usually start at the x axis and because the y axis
is 0 to 499 top-to-bottom.

Also added in some other small changes that seemed to
improve performance--this version also includes
debugging statements, so might be kind of annoying to
run.